Changed directory to tests and extends template base in topic proposal list
