+++ meetupselector/votesystem/templates/topic_proposal_list.html
@@ -1,3 +1,5 @@
+{% extends 'base.html' %}
+{% block content %}
 <table>
   <tr>
     <th>Topic</th>
@@ -16,3 +18,4 @@
   </tr>
   {% endfor %}
 </table>
+{%endblock%}
